// tests/mocks/mock-bucket.repository.d.ts
import { IBucketRepository } from "../../src/domain/repositories/bucket.repository.interface";
export declare class MockBucketRepository implements IBucketRepository {
    getActiveProviders: any;
    listBuckets: any;
    createBucket: any;
    deleteBucket: any;
    setBucketVisibility: any;
    getBucketStats: any;
    listObjects: any;
    uploadFile: any;
    deleteObjects: any;
    createFolder: any;
    searchObjects: any;
    getPresignedUrl: any;
    getObjectStream: any;
    resetAllMocks(): void;
}
//# sourceMappingURL=mock-bucket.repository.d.ts.map